팀 빌딩 알고리즘 작성

+++ team/src/main/java/sansam/team/team/command/application/dto/TeamBuildingDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class TeamBuildingDTO {
     private Long userSeq;
     private Long projectMemberSeq;
-    private int totalScore;
+    private double totalScore;
 
     public TeamBuildingDTO(Long userSeq, Long projectMemberSeq) {
         this.userSeq = userSeq;
